test(2.3): apply test-review findings — split pipeline spec and extract shared factories
Score: 90/100 (A). Applied all 4 findings from bmad-testarch-test-review:

- F-001 HIGH: Split monolithic pipeline.spec.ts (632 lines) into three focused
  files under the 300-line guideline: pipeline-happy-path.spec.ts (happy path
  + parallel fetch + ISR), pipeline-error-handling.spec.ts (failure/partial/
  edge cases), sync-route.spec.ts (CRON_SECRET auth guard). Deleted original.

- F-002 MEDIUM: Extracted duplicate makeRawProperty / makeRawAgent / makeSyncLog
  factories to tests/unit/sync/factories.ts; differ.spec.ts and all pipeline
  specs now import from the shared module.

- F-003 LOW: makeSyncLog() in factories.ts uses FIXED_STARTED_AT constant
  (2026-04-25T12:00:00.000Z) instead of wall-clock new Date(); sync-log.spec.ts
  makeSyncLogRow() extracted matching FIXED_STARTED_AT for consistency.

- F-004 LOW: Route handler imported at module scope in sync-route.spec.ts
  instead of via await import() inside each it() body.

All 77 tests still pass (3 pre-existing skips unchanged).
